+++ packages/ui/src/components/index.ts
@@ -30,6 +30,7 @@ export { PasswordCheck } from './PasswordCheck'
 export { PasswordStrengthMeter } from './PasswordStrengthMeter'
 export { PieChart } from './PieChart'
 export { Popover } from './Popover'
+export { Popup } from './Popup'
 export { ProgressBar } from './ProgressBar'
 export { Radio } from './Radio'
 export { Row } from './Row'
