--- packages/ui/src/internalComponents/index.ts
@@ -1,1 +1,0 @@
-export { Popup } from './Popup'
